% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{RegressionQualityMetric.default}
\alias{RegressionQualityMetric.default}
\title{Get the Default Model Quality RegressionQualityMetric}
\usage{
RegressionQualityMetric.default(x, y)
}
\arguments{
\item{x}{the input coordinates}

\item{y}{the output coordinates}
}
\value{
a default instance of \code{\link{RegressionQualityMetric}} which
  should work well in many scenarios
}
\description{
This method returns a default model quality metric for the given
  coordinates \code{x} and \code{y}. Currently, this is implemented as
  \code{\link{RegressionQualityMetric.sqrtInverseWeightedRmse}}.
}
\seealso{
RegressionQualityMetric.sqrtInverseWeightedRmse
}
