% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionQualityMetric.R
\name{RegressionQualityMetric.new}
\alias{RegressionQualityMetric.new}
\title{Create a new instance of \code{\link{RegressionQualityMetric}}}
\usage{
RegressionQualityMetric.new(quality, x, y, residuals = NULL,
  jacobian = NULL, weights = NULL)
}
\arguments{
\item{quality}{the quality function}

\item{x}{the vector of input values}

\item{y}{the vector if output values}

\item{residuals}{the residuals function (\code{NULL} if undefined)}

\item{jacobian}{the jacobian function (\code{NULL} if undefined)}

\item{weights}{the weights vector (\code{NULL} if undefined)}
}
\value{
the new model quality metric
}
\description{
Instantiate the class \code{\link{RegressionQualityMetric}}
}
