% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{RegressionQualityMetric.rmse}
\alias{RegressionQualityMetric.rmse}
\title{Create an Unweighted Root-Mean Square Error Quality
  RegressionQualityMetric}
\usage{
RegressionQualityMetric.rmse(x, y)
}
\arguments{
\item{x}{the input coordinates}

\item{y}{the output coordinates}
}
\value{
a \code{\link{RegressionQualityMetric}} based on the weighted RMSE
}
\description{
Create a model quality metric based on the \emph{unweighted}
root mean square error. This method is equivalent to calling
\code{\link{RegressionQualityMetric.weightedRmse}} with \code{weights} set
to \code{NULL}.
}
\seealso{
RegressionQualityMetric.new

RegressionQualityMetric.weightedRmse
}
