% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtInverseWeightedRmse.R
\name{RegressionQualityMetric.sqrtInverseWeightedRmse}
\alias{RegressionQualityMetric.sqrtInverseWeightedRmse}
\title{Create a Root-Mean Square Error Quality RegressionQualityMetric where
the Weights are the Inverse of the Square Root of the Absolute \code{y}
Values}
\usage{
RegressionQualityMetric.sqrtInverseWeightedRmse(x, y)
}
\arguments{
\item{x}{the input coordinates}

\item{y}{the output coordinates}
}
\value{
a \code{\link{RegressionQualityMetric}} based on the RMSE weighted by
  the inverse absolute \code{y} values.
}
\description{
Create a model quality metric based on the root mean square
error, using the inverse of the square roots absolute \code{y} values as
weights.
}
\seealso{
RegressionQualityMetric.new

RegressionQualityMetric.weightedRmse

RegressionQualityMetric.inverseWeightedRmse
}
