% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedRmse.R
\name{RegressionQualityMetric.weightedRmse}
\alias{RegressionQualityMetric.weightedRmse}
\title{Create a Weightd Root-Mean Square Error Quality
  RegressionQualityMetric with Given Weights}
\usage{
RegressionQualityMetric.weightedRmse(x, y, weights = NULL)
}
\arguments{
\item{x}{the input coordinates}

\item{y}{the output coordinates}

\item{weights}{\code{NULL} by default, in which case an un-weighted RMSE will
be created}
}
\value{
a \code{\link{RegressionQualityMetric}} based on the weighted RMSE
}
\description{
Create a model quality metric based on the \emph{weighted} root
mean square error.

Weights can be provided via the parameter \code{weights} and are multiplied
to the errors before the squaring. If no weights are provided (or
\code{weights==NULL}), an un-weighted RMSE measure will be created, exactly
as by \code{\link{RegressionQualityMetric.rmse}}.
}
\seealso{
RegressionQualityMetric.new

RegressionQualityMetric.rmse
}
